continued work on facade implementations for JSON and Geotools libraries.
