+++ src/main/java/gov/lanl/micot/application/fragility/Fragility.java
@@ -3,6 +3,7 @@ package gov.lanl.micot.application.fragility;
 import gov.lanl.micot.application.fragility.core.*;
 import gov.lanl.micot.application.fragility.io.GFMDataReader;
 import gov.lanl.micot.application.utilities.asset.PropertyData;
+import gov.lanl.micot.application.utilities.gis.HazardField;
 import gov.lanl.micot.application.utilities.gis.RasterField;
 
 import java.util.ArrayList;
@@ -35,7 +36,7 @@ public class Fragility {
         String[] ids = parser.getIdentifiers();
         String assets = parser.getAssetInputPath();
         GFMDataReader gfmDataReader = new GFMDataReader();
-        ArrayList<RasterField> hazardObjects = gfmDataReader.readHazardFile(hazardFiles, ids);
+        ArrayList<HazardField> hazardObjects = gfmDataReader.readHazardFile(hazardFiles, ids);
 
         // assets
         gfmDataReader.readGeoJsonFile(assets);
