+++ src/main/java/gov/lanl/micot/application/rdt/FragilityRDT.java
@@ -6,6 +6,7 @@ import gov.lanl.micot.application.fragility.core.ResponseEstimator;
 import gov.lanl.micot.application.fragility.core.ResponseEstimatorFactory;
 import gov.lanl.micot.application.fragility.io.GFMDataReader;
 import gov.lanl.micot.application.utilities.asset.PropertyData;
+import gov.lanl.micot.application.utilities.gis.HazardField;
 import gov.lanl.micot.application.utilities.gis.RasterField;
 
 import java.util.ArrayList;
@@ -47,7 +48,7 @@ public class FragilityRDT {
         String[] ids = parser.getIdentifiers();
         String assets = parser.getAssetInputPath();
         GFMDataReader gfmDataReader = new GFMDataReader();
-        ArrayList<RasterField> hazardObjects = gfmDataReader.readHazardFile(hazardFiles, ids);
+        ArrayList<HazardField> hazardObjects = gfmDataReader.readHazardFile(hazardFiles, ids);
 
         // assets
         gfmDataReader.readGeoJsonFile(assets);
