+++ src/main/java/gov/lanl/micot/application/rdt/RDTProcessing.java
@@ -1,31 +1,26 @@
 package gov.lanl.micot.application.rdt;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import gov.lanl.micot.application.fragility.core.*;
+import gov.lanl.micot.application.fragility.core.GFMEngine;
+import gov.lanl.micot.application.fragility.core.GeometryObject;
+import gov.lanl.micot.application.fragility.core.ResponseEstimator;
+import gov.lanl.micot.application.fragility.core.ResponseEstimatorFactory;
 import gov.lanl.micot.application.fragility.io.GFMDataReader;
-import gov.lanl.micot.application.fragility.io.GFMDataWriter;
+import gov.lanl.micot.application.utilities.json.AssetDataFromJackson;
 import gov.lanl.micot.application.utilities.asset.PropertyData;
-import gov.lanl.micot.application.utilities.gis.RasterField;
+import gov.lanl.micot.application.utilities.gis.HazardField;
+import gov.lanl.micot.application.utilities.json.JsonDataFromJackson;
 
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.*;
-
-import static gov.lanl.micot.application.rdt.PoleConstants.*;
 
 /**
  * This class handles RDT data processing and generates the corresponding scenario block
  */
 public final class RDTProcessing {
 
-    private static ObjectMapper mapper = new ObjectMapper();
     private static String scenarioOutputPath = "rdtScenarioBlock.json";
     private static String polesOutputPath = "RDT-to-Poles.json";
     private static int numberOfScenarios = 1;
+
 
     private RDTProcessing() {
     }
@@ -42,8 +37,8 @@ public final class RDTProcessing {
         if (parser.isHasScenarioOutput()) scenarioOutputPath = parser.getScenarioOutput();
         if (parser.isHasRDTPoles()) polesOutputPath = parser.getRdtPolesOutput();
 
-        ObjectNode rdtData = readRDTJSON(filePath);
-        GFMDataWriter.writeResults(rdtData, polesOutputPath);
+        AssetDataFromJackson rdtData = new JsonDataFromJackson();
+        rdtData.readWriteRDTJSON(filePath, polesOutputPath);
 
         String output = parser.getOutputFilePath();
         String estimator = parser.getEstimator();
@@ -53,7 +48,7 @@ public final class RDTProcessing {
         String[] ids = parser.getIdentifiers();
 
         GFMDataReader gfmDataReader = new GFMDataReader();
-        ArrayList<RasterField> hazardObjects = gfmDataReader.readHazardFile(hazardFiles, ids);
+        ArrayList<HazardField> hazardObjects = gfmDataReader.readHazardFile(hazardFiles, ids);
 
         // assets
         gfmDataReader.readGeoJsonFile(polesOutputPath);
@@ -77,16 +72,15 @@ public final class RDTProcessing {
     }
 
     /**
-     * Generates RDT scenario block based-off of response estimations
-     *
-     * @param response
-     * @param assetProperties
+     * Generates RDT scenario block based-off of response estimations.
+     * @param response repsonse values from asset exposures
+     * @param assetProperties asset property data
      */
-    public static void generateScenarios(ArrayNode response, List<Map<String, PropertyData>> assetProperties) {
+    public static void generateScenarios(HashMap<String, Double> response, List<Map<String, PropertyData>> assetProperties) {
 
         Random r = new Random();
 
-        int timeCt = 0;
+        int currentScenarioCount = 0;
 
         HashMap<String, String> poles = new HashMap();
 
@@ -94,200 +88,47 @@ public final class RDTProcessing {
             poles.put(assetProperty.get("id").asString(), assetProperty.get("lineId").asString());
         }
 
-//        // associate pole identifiers to it's JSON node
-//        for (JsonNode n : assetProperties)
-//            poles.put(n.get("id").asText(), n.get("lineId").asText());
-
-
-        ArrayNode scenarioArray = mapper.createArrayNode();
-
-        // using a list to build line ids that are considered damaged.
-        // ArrayNode type has a bug with it's "has" method - produces duplicates
+        // list to build each scenario
         List<String> lineIdList;
+        // global map for all scenarios
+        Map<Integer, List<String>> scenarioLineList = new HashMap<>();
 
         // generating scenarios
         do {
 
             lineIdList = new ArrayList<>();
 
-            for (JsonNode rd : response) {
-                // if response > rand(0,1) --> line is disabled
-                if (rd.get("value").asDouble() > r.nextFloat()) {
-                    String line = poles.get(rd.get("id").asText()).trim();
+            // hasmap iterator
+            for (Map.Entry<String, Double> pair : response.entrySet()){
+                //iterate over the pairs
+                if (pair.getValue() > r.nextFloat()) {
+                    String line = poles.get(pair.getKey()).trim();
 
+                    // ensure no duplicate line numbers add for each scenario
                     if (!lineIdList.contains(line)) {
                         lineIdList.add(line);
                     }
                 }
             }
 
-            timeCt += 1;
-            ArrayNode lineIdentifier = mapper.valueToTree(lineIdList);
+            // add new scenario
+            scenarioLineList.put(currentScenarioCount, new ArrayList<>());
+            scenarioLineList.get(currentScenarioCount).addAll(lineIdList);
 
-            ObjectNode singleScenario = mapper.createObjectNode()
-                    .put("id", String.valueOf(timeCt))
-                    .putPOJO("hardened_disabled_lines", mapper.createArrayNode())
-                    .putPOJO("disable_lines", lineIdentifier);
+            currentScenarioCount += 1;
 
+        } while (currentScenarioCount < numberOfScenarios);
 
-            scenarioArray.add(singleScenario);
-
-        } while (timeCt < numberOfScenarios);
-
-        JsonNode allScenarios = mapper.createObjectNode()
-                .putPOJO("scenarios", scenarioArray);
-
-        GFMDataWriter.writeResults(allScenarios, scenarioOutputPath);
-
+        // write out scenario json file
+        AssetDataFromJackson adj = new JsonDataFromJackson();
+        adj.writeRDTScenarios(scenarioLineList, scenarioOutputPath);
     }
 
-    /**
-     * special method that reads in RDT json and infers pole data from bus locations
-     * @param filePath
-     * @return Jackson GeoJSON formatted object
-     */
-    private static ObjectNode readRDTJSON(String filePath) {
-
-        ObjectMapper mapper = new ObjectMapper();
-        JsonNode fileNodes = null;
-        HashMap<String, List<Double>> nodeLocation;
-        InputStream inStream;
-
-        try {
-            inStream = new FileInputStream(filePath);
-            fileNodes = mapper.readTree(inStream);
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        nodeLocation = new HashMap<>();
-
-        for (JsonNode n : fileNodes.findValue("buses")) {
-            String sid = String.valueOf(n.get("id").asText());
-            double xValue = n.get("x").doubleValue();
-            double yValue = n.get("y").doubleValue();
-
-            nodeLocation.put(sid, new ArrayList<>());
-            nodeLocation.get(sid).add(xValue);
-            nodeLocation.get(sid).add(yValue);
-        }
-
-        int id_count = 0;
-        ArrayNode features = mapper.createArrayNode();
-
-        // create poles for each line segment
-        for (JsonNode n : fileNodes.findValue("lines")) {
-
-            // checking to make sure NOT a transformer
-            boolean xfmrExist = false;
-            JsonNode xmfValue = n.findValue("is_transformer");
-
-            // node may not exist, so checking
-            if (xmfValue != null) {
-                xfmrExist = n.get("is_transformer").asBoolean();
-            }
-
-            if (!xfmrExist) {
-
-                String lid = String.valueOf(n.get("id").asText());
-                String node1 = String.valueOf(n.get("node1_id").asText());
-                String node2 = String.valueOf(n.get("node2_id").asText());
-
-                if (nodeLocation.containsKey(node1) && nodeLocation.containsKey(node2)) {
-                    double x0 = nodeLocation.get(node1).get(0);
-                    double y0 = nodeLocation.get(node1).get(1);
-
-                    // get lat/lon values
-                    double v1 = nodeLocation.get(node2).get(0) - nodeLocation.get(node1).get(0);
-                    double v2 = nodeLocation.get(node2).get(1) - nodeLocation.get(node1).get(1);
-                    double ndist = Math.sqrt(v1 * v1 + v2 * v2);
-
-
-                    // v = (x1,y1) - (x2,y2)
-                    // normalized vector u = v  / ||v||
-                    // now point distance is along line (x1,y1) + du =
-                    v1 = v1 / ndist;
-                    v2 = v2 / ndist;
-
-                    // using 111.111 km per degree (or 111,111 meters)
-                    // pole spacing at 91 meters 0.000817463169242 degrees
-                    // 0.000817463169242
-                    double numPoles = Math.floor(ndist * DEG_TO_METERS) / POLE_SPACING;
-                    numPoles = Math.floor(numPoles);
-
-                    ObjectNode obj;
-                    for (int i = 0; i < numPoles; i++) {
-                        obj = createPoleAsset(id_count, lid, new double[]{x0, y0});
-
-                        features.add(obj);
-
-                        // move to next pole location
-                        x0 = x0 + v1 * NEXT_DISTANCE;
-                        y0 = y0 + v2 * NEXT_DISTANCE;
-                        id_count = id_count + 1;
-                    }
-                } else {
-                    System.out.println(node1 + " and/or " + node2 + " doesn't exist for line id " + lid + "");
-                }
-            }
-
-        }// end for
-
-
-        System.out.println("Number of poles created: " + id_count);
-
-        // create geoJSON Feature Collection
-        ObjectNode featureNode = mapper.createObjectNode()
-                .put("type", "FeatureCollection")
-                .putPOJO("features", features);
-
-        return featureNode;
-    }
-
-    private static ObjectNode createPoleAsset(int id, String line, double[] coord) {
-
-        String sid = String.valueOf(id);
-        // creates one GeoJSON Point Object
-        ObjectNode featureNode = mapper.createObjectNode()
-                .putPOJO("geometry", mapper.createObjectNode()
-                        .put("type", "Point")
-                        .putPOJO("coordinates", mapper.createArrayNode()
-                                .add(coord[0])
-                                .add(coord[1])
-                        ))
-                .put("type", "Feature")
-
-                .putPOJO("properties", mapper.createObjectNode()
-                        .put("id", sid)
-                        .put("baseDiameter", BASE_DIAMETER)
-                        .put("cableSpan", CABLE_SPAN)
-                        .put("commAttachmentHeight", COMM_ATTACHMENT_HEIGHT)
-                        .put("commCableDiameter", COMM_CABLE_DIAMETER)
-                        .put("commCableNumber", COMM_CABLE_NUMBER)
-                        .put("commCableWireDensity", COMM_CABLE_WIRE_DENSITY)
-                        .put("height", HEIGHT)
-                        .put("meanPoleStrength", MEAN_POLE_STRENGTH)
-                        .put("powerAttachmentHeight", POWER_ATTACHMENT_HEIGHT)
-                        .put("powerCableDiameter", POWER_CABLE_DIAMETER)
-                        .put("powerCableNumber", POWER_CABLE_NUMBER)
-                        .put("powerCableWireDensity", POWER_CABLE_WIRE_DENSITY)
-                        .put("powerCircuitName", POWER_CIRCUIT_NAME)
-                        .put("stdDevPoleStrength", STD_DEV_POLE_STRENGTH)
-                        .put("topDiameter", TOP_DIAMETER)
-                        .put("woodDensity", WOOD_DENSITY)
-                        .put("lineId", line));
-
-        return featureNode;
-
-    }
-
-    public static void setNumberOfScenarios(int number){
+    static void setNumberOfScenarios(int number){
         numberOfScenarios = number;
     }
 
-    public static void setScenarioOutputPath(String path){
+    static void setScenarioOutputPath(String path){
         scenarioOutputPath = path;
     }
-
 }
