--- src/main/java/gov/lanl/micot/application/utilities/asset/AssetString.java
@@ -1,4 +1,0 @@
-package gov.lanl.micot.application.utilities.asset;
-
-public class AssetString  {
-}
