+++ src/main/java/gov/lanl/micot/application/utilities/asset/StringValue.java
@@ -22,4 +22,9 @@ public class StringValue extends ValueTypes {
     public boolean isString(){
         return true;
     }
+
+    @Override
+    public boolean isBoolean() {
+        return false;
+    }
 }
