+++ src/main/java/gov/lanl/micot/application/utilities/asset/ValueTypes.java
@@ -1,11 +1,14 @@
 package gov.lanl.micot.application.utilities.asset;
 
 public abstract class ValueTypes extends PropertyData {
+
     public abstract boolean booleanValue();
 
     public abstract String StringValue();
 
     public abstract boolean isString();
+
+    public abstract boolean isBoolean();
 
     protected ValueTypes() {
     }
@@ -19,6 +22,10 @@ public abstract class ValueTypes extends PropertyData {
     }
 
     public final String asString() {
+        return this.StringValue();
+    }
+
+    public final String asString(String defaultValue) {
         return this.StringValue();
     }
