+++ src/main/java/gov/lanl/micot/application/utilities/gis/geotools/GeotoolsGIS.java
@@ -4,6 +4,4 @@ package gov.lanl.micot.application.utilities.gis.geotools;
  * @author Trevor Crawford
  */
 public interface GeotoolsGIS {
-
-
 }
