+++ src/test/java/gov/lanl/micot/application/fragility/io/GFMDataReaderTest.java
@@ -3,6 +3,7 @@ package gov.lanl.micot.application.fragility.io;
 import com.fasterxml.jackson.databind.JsonNode;
 import gov.lanl.micot.application.fragility.core.GeometryObject;
 import gov.lanl.micot.application.utilities.asset.PropertyData;
+import gov.lanl.micot.application.utilities.gis.HazardField;
 import gov.lanl.micot.application.utilities.gis.RasterField;
 import junit.framework.TestCase;
 
@@ -21,7 +22,7 @@ public class GFMDataReaderTest extends TestCase {
         String[] ip = new String[1];
         ip[0] = "ice";
         fp[0] = "test_data/fields/windField_example.asc";
-        ArrayList<RasterField> hazardObjects1 = gfmdr.readHazardFile(fp,ip);
+        ArrayList<HazardField> hazardObjects1 = gfmdr.readHazardFile(fp,ip);
         assertTrue(!hazardObjects1.isEmpty());
 
     }
@@ -33,7 +34,7 @@ public class GFMDataReaderTest extends TestCase {
         String[] ip = new String[1];
         ip[0] = "ice";
         fp[0] = "test_data/fields/windField_example.tif";
-        ArrayList<RasterField> hazardObjects1 = gfmdr.readHazardFile(fp,ip);
+        ArrayList<HazardField> hazardObjects1 = gfmdr.readHazardFile(fp,ip);
         assertTrue(!hazardObjects1.isEmpty());
 
     }
@@ -49,7 +50,7 @@ public class GFMDataReaderTest extends TestCase {
         fp1[0] = "test_data/fields/iceField_example.asc";
         fp1[1] = "test_data/fields/windField_example.asc";
 
-        ArrayList<RasterField> hazardObjects2 = gfmdr.readHazardFile(fp1,ip1);
+        ArrayList<HazardField> hazardObjects2 = gfmdr.readHazardFile(fp1,ip1);
         assertTrue(!hazardObjects2.isEmpty());
 
     }
