+++ src/test/java/gov/lanl/micot/application/fragility/responseEstimators/PowerPoleWindIceStressTest.java
@@ -5,6 +5,7 @@ import gov.lanl.micot.application.fragility.FragilityParameters;
 import gov.lanl.micot.application.fragility.core.*;
 import gov.lanl.micot.application.fragility.io.GFMDataReader;
 import gov.lanl.micot.application.utilities.asset.PropertyData;
+import gov.lanl.micot.application.utilities.gis.HazardField;
 import gov.lanl.micot.application.utilities.gis.RasterField;
 import junit.framework.TestCase;
 
@@ -47,7 +48,7 @@ public class PowerPoleWindIceStressTest extends TestCase {
 
         GFMDataReader gfmdr = new GFMDataReader();
 
-        ArrayList<RasterField> hazardObjects = gfmdr.readHazardFile(hazardFiles, ids);
+        ArrayList<HazardField> hazardObjects = gfmdr.readHazardFile(hazardFiles, ids);
 
         // assets
         gfmdr.readGeoJsonFile(assets);
