+++ src/test/java/gov/lanl/micot/application/rdt/RDTExamplePolesTests.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import gov.lanl.micot.application.fragility.core.*;
 import gov.lanl.micot.application.fragility.io.GFMDataReader;
 import gov.lanl.micot.application.utilities.asset.PropertyData;
+import gov.lanl.micot.application.utilities.gis.HazardField;
 import gov.lanl.micot.application.utilities.gis.RasterField;
 import junit.framework.TestCase;
 
@@ -137,7 +138,7 @@ public class RDTExamplePolesTests extends TestCase {
         String[] ids = parser.getIdentifiers();
         String assets = parser.getAssetInputPath();
         GFMDataReader gfmdr = new GFMDataReader();
-        ArrayList<RasterField> hazardObjects = gfmdr.readHazardFile(hazardFiles, ids);
+        ArrayList<HazardField> hazardObjects = gfmdr.readHazardFile(hazardFiles, ids);
 
         // assets
         gfmdr.readGeoJsonFile(assets);
